Restructure writers + instance sharing + closing
The base class for event writer implementations
(`ElasticSearchWriteAccess`) combines both the configuration and the
execution. This leads to ugly classes with too much functionality.

Replace `ElasticSearchWriteAccess` by three separate types with clear
focus:

- `EventWriterGlobalConfig`: The global configuration of an event writer
  implementation, which is managed by Jenkins.

- `EventWriterRunConfiguration`: An immutable serializable and remotable
  configuration snapshot that gets part of an
  `ElasticSearchRunConfiguration`.

- `EventWriter`: The execution.

Secondly, ensure that implementations of `EventWriterRunConfig` create
_shared_ instances of `EventWriter`. This minimizes resource usage,
increases performance and avoids multiple independent transmissions of
events of the same run.

Third, ensure that all `EventWriter` instances are properly closed to
avoid resource leaks.
