+++ src/main/java/io/jenkins/plugins/pipeline_elasticsearch_logs/ElasticSearchRunListener.java
@@ -1,7 +1,6 @@
 package io.jenkins.plugins.pipeline_elasticsearch_logs;
 
-import java.io.IOException;
-import java.net.URISyntaxException;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -10,22 +9,52 @@ import hudson.Extension;
 import hudson.model.Result;
 import hudson.model.Run;
 import hudson.model.listeners.RunListener;
-import io.jenkins.plugins.pipeline_elasticsearch_logs.write.ElasticSearchWriteAccess;
+import io.jenkins.plugins.pipeline_elasticsearch_logs.utils.RunUtils;
+import io.jenkins.plugins.pipeline_elasticsearch_logs.write.EventWriter;
 
 @Extension
 public class ElasticSearchRunListener extends RunListener<Run<?, ?>> {
 
+    private static final Logger LOGGER = Logger.getLogger(ElasticSearchRunListener.class.getName());
+
+    private final Map<String, EventWriter> eventWritersByRunId = new HashMap<>();
+
     @Override
-    public void onFinalized(Run<?, ?> run) {
+    public void onInitialize(Run<?, ?> run) {
+
         try {
             ElasticSearchRunConfiguration config = ElasticSearchRunConfiguration.get(run);
             if (config == null) {
                 return;
             }
 
-            ElasticSearchWriteAccess writer = config.createWriteAccess();
-            Map<String, Object> data = config.createData();
+            EventWriter eventWriter = config.createEventWriter();
+            this.eventWritersByRunId.put(RunUtils.getUniqueRunId(run), eventWriter);
 
+            Map<String, Object> data = config.createData();
+            data.put("eventType", "buildStart");
+            eventWriter.push(data);
+        }
+        catch (Exception e) {
+            LOGGER.log(Level.SEVERE, "Failed to get Executable of FlowExecution.", e);
+        }
+    }
+
+    @Override
+    public void onFinalized(Run<?, ?> run) {
+        EventWriter writer = null;
+        try {
+            ElasticSearchRunConfiguration config = ElasticSearchRunConfiguration.get(run);
+            if (config == null) {
+                return;
+            }
+            writer = this.eventWritersByRunId.remove(RunUtils.getUniqueRunId(run));
+            if (writer == null) {
+                LOGGER.log(Level.SEVERE, "internal inconsistency: Elasticsearch event writer not found although it should have been created at the start of the run");
+                return;
+            }
+
+            Map<String, Object> data = config.createData();
             data.put("eventType", "buildEnd");
             Result result = run.getResult();
             if (result != null) {
@@ -37,33 +66,19 @@ public class ElasticSearchRunListener extends RunListener<Run<?, ?>> {
             }
             writer.push(data);
         }
-        catch (IOException | URISyntaxException e) {
+        catch (Exception e) {
             LOGGER.log(Level.SEVERE, "Failed to get Executable of FlowExecution.");
         }
         finally {
-            ElasticSearchRunConfiguration.release(run);
-        }
-    }
-
-    private static final Logger LOGGER = Logger.getLogger(ElasticSearchRunListener.class.getName());
-
-    @Override
-    public void onInitialize(Run<?, ?> run) {
-
-        try {
-            ElasticSearchRunConfiguration config = ElasticSearchRunConfiguration.get(run);
-            if (config == null) {
-                return;
+            if (writer != null) {
+                try {
+                    writer.close();
+                }
+                catch (Exception ex) {
+                    LOGGER.log(Level.SEVERE, "failed to close event writer", ex);
+                }
             }
-
-            ElasticSearchWriteAccess writer = config.createWriteAccess();
-            Map<String, Object> data = config.createData();
-
-            data.put("eventType", "buildStart");
-            writer.push(data);
-        }
-        catch (IOException | URISyntaxException e) {
-            LOGGER.log(Level.SEVERE, "Failed to get Executable of FlowExecution.", e);
+            ElasticSearchRunConfiguration.release(run);
         }
     }
 }
