--- src/main/java/io/jenkins/plugins/pipeline_elasticsearch_logs/SerializableSupplier.java
@@ -1,9 +1,0 @@
-package io.jenkins.plugins.pipeline_elasticsearch_logs;
-
-
-import java.io.Serializable;
-import java.util.function.Supplier;
-
-public interface SerializableSupplier<T> extends Supplier<T>, Serializable {
-    // Only method inherited from Supplier
-}
