+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdEventWriterGlobalConfig/help-bufferChunkRetentionTimeMillis.html
@@ -1,0 +1,3 @@
+<div>
+The time (milliseconds) after which data in the buffer will be flushed latest.
+</div>
