+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdEventWriterGlobalConfig/help-connectionTimeoutMillis.html
@@ -1,0 +1,5 @@
+<div>
+  The timeout in milliseconds for establishing a connection to the destination
+  server.
+  If empty, a default value will be used.
+</div>
