+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdEventWriterGlobalConfig/help-emitMaxRetriesIfBufferFull.html
@@ -1,0 +1,6 @@
+<div>
+The maximum number of retries when inserting a log event into the in-memory buffer
+fails because the buffer is full.
+A negative value means to retry forever.
+If empty, a default value will be used.
+</div>
