+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdEventWriterGlobalConfig/help-host.html
@@ -1,0 +1,3 @@
+<div>
+The host name of the destination server.
+</div>
