+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdEventWriterGlobalConfig/help-readTimeoutMillis.html
@@ -1,0 +1,5 @@
+<div>
+  The timeout in milliseconds for waiting for the acknowledgement from
+  the destination server after sending a chunk of events.
+  If empty, a default value will be used.
+</div>
