+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdEventWriterGlobalConfig/help.html
