--- src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdWriter/help-bufferRetentionMillis.html
@@ -1,3 +1,0 @@
-<div>
-The time in milliseconds after which data in the buffer will be flushed latest. 
-</div>
