--- src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdWriter/help-host.html
@@ -1,3 +1,0 @@
-<div>
-Host name of the fluentd instance to use.
-</div>
