--- src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdWriter/help-maxRetries.html
@@ -1,3 +1,0 @@
-<div>
-The maximum number of times the sender should try to resend an event in case of an error. 
-</div>
