--- src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/fluentd/FluentdWriter/help-timeoutMillis.html
@@ -1,3 +1,0 @@
-<div>
-Timeout in milliseconds the sender waits before failing on connection issues.
-</div>
