+++ src/main/resources/io/jenkins/plugins/pipeline_elasticsearch_logs/write/index_api/IndexAPIEventWriterGlobalConfig/help-indexUrl.html
